Fix clippy and fmt warnings
